Fix two CI failures a Mac build could not see
The firmware host suite failed only under GCC: comparing constants from two
different anonymous enums is -Werror=enum-compare there and silent under
Apple clang, so face_sprite_sheet_test.c built locally and broke CI. Cast
both sides to int. A sweep for the same pattern found no other
first-party instance.

The preview lane failed for a subtler reason, and the fix is a guard this
file already contains elsewhere. GitHub's compare response caps its file
list at 300, and the list is ORDERED — so on a large diff the tail is
simply absent, and absence reads as "that app did not change". This branch
has 386 changed files whose `apps/kit/**` paths sort before `apps/os/**`:
the firmware consumed the entire cap, OS was never selected, nothing
deployed, and the e2e lane then had no recorded deployment to test and
failed the PR. The container-rollout decision already refuses to trust a
capped comparison for exactly this reason; app selection now refuses the
same way, and the only safe refusal is the whole fleet.

Measured against the real function: 300 changed files select all 6 apps,
299 still select none. Note that scripts/preview/preview.test.ts — where
these two cases are asserted — does not appear to be included by any test
lane, so the assertions were verified by driving the exported internals
directly. That gap is pre-existing and worth its own fix.

+++ apps/kit/firmware/components/avatar/tests/face_sprite_sheet_test.c
@@ -708,7 +708,10 @@ int main(void)
 {
     assert(sizeof(face_render_key_t) == FACE_RENDER_KEY_BYTES);
     assert(FACE_RENDER_KEY_BYTES == 40);
-    assert(TEST_MOUTH_SLOT_COUNT > FACE_VISEME_COUNT);
+    /* Cast to int: GCC's -Werror=enum-compare rejects comparing constants
+     * from two different anonymous enums, and clang does not — so this line
+     * built on a Mac and broke CI. */
+    assert((int)TEST_MOUTH_SLOT_COUNT > (int)FACE_VISEME_COUNT);
 
     test_validation();
     test_vocabularies_and_fallback();
